requirements.txt
# Core Framework
fastapi>=0.95.0
flask>=2.2.3
uvicorn>=0.21.1

# UI Framework
gradio>=3.23.0

# Database
sqlalchemy>=2.0.7
alembic>=1.10.2

# Authentication
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1

# File Processing
pydub>=0.25.1
librosa>=0.10.0
soundfile>=0.12.1

# Resemble AI Integration
requests>=2.28.2
websockets>=10.4
aiohttp>=3.8.4

# Utilities
python-dotenv>=1.0.0
pydantic>=1.10.7
tqdm>=4.65.0
loguru>=0.6.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0

# Development
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
